+++ src/main/java/com/anthonyhilyard/iceberg/IcebergClient.java
@@ -1,6 +1,14 @@
 package com.anthonyhilyard.iceberg;
 
+import com.anthonyhilyard.iceberg.util.Tooltips.TitleBreakComponent;
+import com.mojang.datafixers.util.Either;
+
+import net.minecraft.network.chat.FormattedText;
+import net.minecraft.world.inventory.tooltip.TooltipComponent;
 import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.client.event.RenderTooltipEvent.GatherComponents;
+import net.minecraftforge.eventbus.api.EventPriority;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventBusSubscriber.Bus;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
@@ -14,7 +22,23 @@ public class IcebergClient
 
 	public void onClientSetup(FMLClientSetupEvent event)
 	{
-	
+	}
+
+	@SubscribeEvent(priority = EventPriority.LOWEST)
+	public static void onGatherComponentsEventEnd(GatherComponents event)
+	{
+		if (event.getTooltipElements().size() > 1)
+		{
+			// Insert a title break component after the first formattedText component.
+			for (int i = 0; i < event.getTooltipElements().size(); i++)
+			{
+				if (event.getTooltipElements().get(i).left().isPresent())
+				{
+					event.getTooltipElements().add(i + 1, Either.<FormattedText, TooltipComponent>right(new TitleBreakComponent()));
+					break;
+				}
+			}
+		}
 	}
 
 	// @SubscribeEvent